requirements.txt
requests>=2.28.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
joblib>=1.2.0
matplotlib>=3.5.0
seaborn>=0.11.0
flask>=2.2.0
streamlit>=1.28.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
plotly>=5.15.0
